Fixed #34: Make it possible to double tap when placing items
This fixes having selected a button or light and trying to double tap a component, which will place the light and then prevent any other action

It does this by having a grace period where the state can be reversed. This happens if a double-tap action is performed less than 500ms after a tap-tile action.